Move `Region.toSourceRange(...)` to `Region.swift`
