+++ Source/SwiftLintFramework/Rules/Style/ClosureSpacingRule.swift
@@ -262,20 +262,6 @@ private extension TokenSyntax {
     }
 }
 
-private extension Region {
-    func toSourceRange(locationConverter: SourceLocationConverter) -> SourceRange? {
-        guard let startLine = start.line, let endLine = end.line else {
-            return nil
-        }
-
-        let startPosition = locationConverter.position(ofLine: startLine, column: min(1000, start.character ?? 1))
-        let endPosition = locationConverter.position(ofLine: endLine, column: min(1000, end.character ?? 1))
-        let startLocation = locationConverter.location(for: startPosition)
-        let endLocation = locationConverter.location(for: endPosition)
-        return SourceRange(start: startLocation, end: endLocation)
-    }
-}
-
 private extension SourceRange {
     func contains(_ position: AbsolutePosition, locationConverter: SourceLocationConverter) -> Bool {
         contains(locationConverter.location(for: position))
